PR review: - Removed "linear" folder and moved FourierLayer into fourier_reparametrized.py - Simplified model creation. - Added additional test for nonlinearity error.
